Move classloading utils from separate module to core.internal
