--- common/classloading/README.md
@@ -1,3 +1,0 @@
-# classloading
-
-This module provides common functions around classloading
