--- common/classloading/build.gradle
@@ -1,23 +1,0 @@
-apply plugin: 'kotlin'
-
-apply plugin: 'net.corda.plugins.publish-utils'
-apply plugin: 'com.jfrog.artifactory'
-
-dependencies {
-    compile group: "org.jetbrains.kotlin", name: "kotlin-stdlib-jdk8", version: kotlin_version
-    compile group: "org.jetbrains.kotlin", name: "kotlin-reflect", version: kotlin_version
-
-    compile group: "io.github.classgraph", name: "classgraph", version: class_graph_version
-
-    testCompile group: "org.jetbrains.kotlin", name: "kotlin-test", version: kotlin_version
-    testCompile group: "junit", name: "junit", version: junit_version
-    testCompile group: "org.assertj", name: "assertj-core", version: assertj_version
-}
-
-jar {
-    baseName 'corda-common-classloading'
-}
-
-publish {
-    name jar.baseName
-}
