+++ core/build.gradle
@@ -112,7 +112,7 @@ dependencies {
     // required to use @Type annotation
     compile "org.hibernate:hibernate-core:$hibernate_version"
 
-    compile project(":common-classloading")
+    compile group: "io.github.classgraph", name: "classgraph", version: class_graph_version
 }
 
 // TODO Consider moving it to quasar-utils in the future (introduced with PR-1388)
